Pass data from admin page to the dynamic route
