+++ client/src/views/Admin.vue
@@ -56,419 +56,16 @@
 
           <v-card-text>
             <v-tabs-window v-model="activeTab">
-              <router-view></router-view>
+              <router-view
+                :vouchers="vouchers"
+                :pendingVouchers="pendingVouchers"
+                @update-vouchers="getVouchers"
+              />
             </v-tabs-window>
           </v-card-text>
         </v-card>
       </v-col>
     </v-row>
-    <!-- <v-row>
-      <v-col cols="12" md="8">
-        <v-data-table
-          v-model:items-per-page="itemsPerPage"
-          :headers="vouchersHeaders"
-          :items="vouchers"
-          class="elevation-1 my-2 rounded shadow"
-        >
-          <template v-slot:item="{ item }">
-            <tr>
-              <td>{{ item.id }}</td>
-              <td v-if="item.name" class="d-flex align-center">
-                <v-avatar color="primary" size="30" class="mr-2">
-                  <span class="text-uppercase">{{ addAvatar(item.name) }}</span>
-                </v-avatar>
-                <div>
-                  <p>{{ item.name }}</p>
-                  <p>
-                    {{ item.email }}
-                  </p>
-                </div>
-              </td>
-
-              <td class="d-flex align-center" v-else>
-                <v-avatar color="warning" size="30" class="mr-2">
-                  <span class="text-uppercase">A</span>
-                </v-avatar>
-                <div>
-                  <p>Voucher generated by Admin</p>
-                </div>
-              </td>
-              <td v-if="!dateIsNull(item.updated_at)">
-                {{ stringDate(item.updated_at) }}
-              </td>
-              <td v-else>-</td>
-              <td v-if="item.reason">{{ item.reason }}</td>
-              <td v-else>-</td>
-              <td>{{ item.vms }}</td>
-              <td>{{ item.public_ips }}</td>
-              <td>{{ item.voucher }}</td>
-              <td v-if="item.rejected">
-                <span class="text-error">Rejected</span>
-              </td>
-              <td v-else>
-                <span class="text-success">Approved</span>
-              </td>
-            </tr>
-          </template>
-        </v-data-table>
-        <v-data-table
-          v-model:items-per-page="itemsPerPage"
-          :headers="pendingVouchersHeaders"
-          :items="pendingVouchers"
-          class="elevation-1 my-2 rounded shadow"
-        >
-          <template v-slot:item="{ item }">
-            <tr>
-              <td>{{ item.id }}</td>
-              <td v-if="item.name" class="d-flex align-center">
-                <v-avatar color="primary" size="30" class="mr-2">
-                  <span class="text-uppercase">{{ addAvatar(item.name) }}</span>
-                </v-avatar>
-                <div>
-                  <p>{{ item.name }}</p>
-                  <p>
-                    {{ item.email }}
-                  </p>
-                </div>
-              </td>
-
-              <td class="d-flex align-center" v-else>
-                <v-avatar color="warning" size="30" class="mr-2">
-                  <span class="text-uppercase">A</span>
-                </v-avatar>
-                <div>
-                  <p>Voucher generated by Admin</p>
-                </div>
-              </td>
-              <td v-if="!dateIsNull(item.created_at)">
-                {{ stringDate(item.created_at) }}
-              </td>
-              <td v-else>-</td>
-              <td v-if="item.reason">{{ item.reason }}</td>
-              <td v-else>-</td>
-              <td>{{ item.vms }}</td>
-              <td>{{ item.public_ips }}</td>
-              <td v-if="!item.approved && !item.rejected">
-                <v-row>
-                  <v-col cols="6">
-                    <BaseButton
-                      color="primary"
-                      text="Approve"
-                      size="small"
-                      class="mr-1 px-8 text-capitalize"
-                      variant="flat"
-                      @click="approveVoucher(item.id, true)"
-                    />
-                  </v-col>
-                  <v-col cols="6">
-                    <BaseButton
-                      color="red-lighten-1"
-                      text="Reject"
-                      size="small"
-                      class="text-capitalize"
-                      variant="flat"
-                      @click="approveVoucher(item.id, false)"
-                    />
-                  </v-col>
-                </v-row>
-              </td>
-            </tr>
-          </template>
-        </v-data-table>
-        <div class="actions d-flex justify-end my-2">
-          <div class="text-center">
-            <v-dialog v-model="announcementDialog" width="auto">
-              <template v-slot:activator="{ props }">
-                <BaseButton
-                  color="primary"
-                  text="Send an announcement"
-                  class="text-capitalize mr-2"
-                  v-bind="props"
-                />
-              </template>
-
-              <v-card class="pa-5">
-                <v-form @submit.prevent="sendAnnouncement" ref="form">
-                  <v-card-text>
-                    <h5
-                      class="text-h5 text-md-h4 text-center mb-5 pa-5 secondary"
-                    >
-                      Send an announcement
-                    </h5>
-                    <v-text-field
-                      label="Subject"
-                      v-model="subject"
-                      :rules="requiredRules"
-                      oninput="validity.valid||(value='')"
-                      bg-color="accent"
-                      variant="outlined"
-                      density="compact"
-                      class="my-3"
-                    ></v-text-field>
-
-                    <v-textarea
-                      clearable
-                      label="Announcement"
-                      v-model="announcement"
-                      :rules="requiredRules"
-                      oninput="validity.valid||(value='')"
-                      bg-color="accent"
-                      variant="outlined"
-                      density="compact"
-                      class="my-3"
-                    ></v-textarea>
-                  </v-card-text>
-                  <v-card-actions class="justify-center">
-                    <BaseButton
-                      class="bg-primary mr-5"
-                      text="Cancel"
-                      @click="announcementDialog = false"
-                    />
-                    <BaseButton type="submit" class="bg-primary" text="Send" />
-                  </v-card-actions>
-                </v-form>
-              </v-card>
-            </v-dialog>
-          </div>
-          <div class="text-center">
-            <v-dialog v-model="dialog" width="auto">
-              <template v-slot:activator="{ props }">
-                <BaseButton
-                  color="warning"
-                  text="Generate a voucher"
-                  class="text-capitalize mr-2"
-                  v-bind="props"
-                />
-              </template>
-
-              <v-card class="pa-5">
-                <v-form @submit.prevent="generateVoucher" ref="form">
-                  <v-card-text>
-                    <h5
-                      class="text-h5 text-md-h4 text-center mb-5 pa-5 secondary"
-                    >
-                      Generate a voucher
-                    </h5>
-                    <v-row>
-                      <v-col>
-                        <v-text-field
-                          label="VMs"
-                          v-model="vms"
-                          :rules="requiredRules"
-                          min="1"
-                          type="number"
-                          oninput="validity.valid||(value='')"
-                          bg-color="accent"
-                          variant="outlined"
-                          density="compact"
-                        ></v-text-field>
-                      </v-col>
-                      <v-col>
-                        <v-text-field
-                          label="IPs"
-                          v-model="ips"
-                          :rules="requiredRules"
-                          min="0"
-                          oninput="validity.valid||(value='')"
-                          type="number"
-                          bg-color="accent"
-                          variant="outlined"
-                          density="compact"
-                        ></v-text-field>
-                      </v-col>
-                    </v-row>
-
-                    <v-text-field
-                      label="Length"
-                      v-model="length"
-                      :rules="requiredRules"
-                      min="3"
-                      max="20"
-                      oninput="validity.valid||(value='')"
-                      type="number"
-                      bg-color="accent"
-                      variant="outlined"
-                      density="compact"
-                      class="my-3"
-                    ></v-text-field>
-                  </v-card-text>
-                  <v-card-actions class="justify-center">
-                    <BaseButton
-                      class="bg-primary mr-5"
-                      text="Cancel"
-                      @click="dialog = false"
-                    />
-                    <BaseButton
-                      type="submit"
-                      class="bg-primary"
-                      text="Generate"
-                    />
-                  </v-card-actions>
-                </v-form>
-              </v-card>
-            </v-dialog>
-          </div>
-          <div class="text-center">
-            <template v-if="nextLaunchDialog">
-              <BaseButton
-                color="green"
-                text="Disable Launch"
-                class="text-capitalize mr-2"
-                @click="setNextLaunch"
-              />
-            </template>
-            <template v-else>
-              <BaseButton
-                color="red"
-                text="Enable Launch"
-                class="text-capitalize mr-2"
-                @click="setNextLaunch"
-              />
-            </template>
-          </div>
-          <BaseButton
-            :disabled="approveAllCount <= 0"
-            color="success"
-            text="Approve All"
-            class="approve text-capitalize"
-            @click="approveAllVouchers"
-          />
-        </div>
-      </v-col>
-      <v-col cols="12" md="4">
-        <v-col class="pa-0">
-          <div
-            class="balance text-primary pa-3 my-2 text-center rounded-lg bg-white shadow"
-          >
-            <p class="mx-lg-auto font-weight-medium">Balance</p>
-            <strong style="font-size: 2rem">{{ balance }} TFT</strong>
-          </div>
-        </v-col>
-        <v-row>
-          <v-col>
-            <div
-              class="resources text-white text-center rounded-lg bg-primary py-5 shadow"
-            >
-              <p class="mx-lg-auto font-weight-medium">
-                Used VMs: {{ usedResources }}
-              </p>
-              <p class="mx-lg-auto font-weight-medium">
-                Deployed VMs: {{ deployedResources }}
-              </p>
-            </div>
-          </v-col>
-          <v-col>
-            <div
-              class="resources text-white text-center rounded-lg bg-primary py-5 shadow"
-            >
-              <p class="mx-lg-auto font-weight-medium">
-                Used IPs: {{ usedIPs }}
-              </p>
-              <p class="mx-lg-auto font-weight-medium">
-                Reserved IPs: {{ reservedIPs }}
-              </p>
-            </div>
-          </v-col>
-        </v-row>
-        <section class="my-5">
-          <v-data-table
-            v-model:items-per-page="itemsPerPage"
-            :headers="usersHeaders"
-            :items="users"
-            class="elevation-1"
-          >
-            <template v-slot:item="{ item, index }">
-              <tr>
-                <td>
-                  {{ ++index }}
-                </td>
-                <td v-if="item.name" class="d-flex align-center">
-                  <v-avatar color="primary" size="30" class="mr-2">
-                    <v-icon
-                      v-if="item.admin"
-                      icon="mdi-account-key"
-                      size="17"
-                    ></v-icon>
-                    <span v-else class="text-uppercase">
-                      {{ addAvatar(item.name) }}
-                    </span>
-                  </v-avatar>
-                  <div>
-                    <p>{{ item.name }}</p>
-                    <p>
-                      {{ item.email }}
-                    </p>
-                  </div>
-                </td>
-                <td>
-                  <span class="text-red">{{ item.used_vms }}</span>
-                  /<span>{{ item.vms }}</span>
-                </td>
-                <td>
-                  <span class="text-red">{{ item.used_public_ips }}</span
-                  >/<span>{{ item.public_ips }}</span>
-                </td>
-                <td>
-                  <v-row>
-                    <v-tooltip block text="View user" left>
-                      <template v-slot:activator="{ props }">
-                        <v-icon
-                          v-bind="props"
-                          color="primary"
-                          dark
-                          class="ml-3 text-primary cursor-pointer"
-                          @click="openUserInfo(item)"
-                        >
-                          mdi-information
-                        </v-icon>
-                      </template>
-                    </v-tooltip>
-
-                    <v-tooltip v-if="!item.admin" block text="Set admin" left>
-                      <template v-slot:activator="{ props }">
-                        <v-icon
-                          v-bind="props"
-                          color="primary"
-                          dark
-                          class="ml-1 text-primary cursor-pointer"
-                          @click="setAdmin(item, true)"
-                        >
-                          mdi-account-key
-                        </v-icon>
-                      </template>
-                    </v-tooltip>
-
-                    <v-tooltip v-if="item.admin" block text="Remove admin" left>
-                      <template v-slot:activator="{ props }">
-                        <v-icon
-                          v-bind="props"
-                          color="primary"
-                          dark
-                          class="ml-1 text-primary cursor-pointer"
-                          @click="setAdmin(item, false)"
-                        >
-                          mdi-account-lock
-                        </v-icon>
-                      </template>
-                    </v-tooltip>
-                  </v-row>
-                </td>
-              </tr>
-            </template>
-          </v-data-table>
-        </section>
-      </v-col>
-    </v-row>
-    <v-dialog transition="dialog-top-transition" v-model="showUserInfo">
-      <UserInfo :user="userInfo"></UserInfo>
-    </v-dialog>
-    <Toast ref="toast" />
-    <Voucher
-      v-if="voucher"
-      :msg="message"
-      :voucher="voucher"
-      :reset="resetVoucher"
-    /> -->
     <Toast ref="toast" />
   </v-container>
 </template>
@@ -480,43 +77,12 @@ import userService from "@/services/userService.js";
 import Toast from "@/components/Toast.vue";
 import { storeToRefs } from "pinia";
 import { useUserStore } from "@/store/UserStore";
-// import Voucher from "@/components/Voucher.vue";
-// import UserInfo from "@/components/UserInfo.vue";
-// import { useRouter } from "vue-router";
+import router from "@/router";
 
-// const router = useRouter();
-// const confirm = ref(null);
-// const vouchersHeaders = ref([
-//   { title: "No", key: "id" },
-//   { title: "User", key: "user", sortable: false },
-//   { title: "Updated at", key: "updated_at" },
-//   { title: "Reason for Voucher", key: "reason", sortable: false },
-//   { title: "VMs", key: "vms" },
-//   { title: "Public IPs", key: "public_ips" },
-//   { title: "Voucher", key: "voucher" },
-//   { title: "Actions", key: "actions", sortable: false },
-// ]);
-
-// const pendingVouchersHeaders = ref([
-//   { title: "No", key: "id" },
-//   { title: "User", key: "user", sortable: false },
-//   { title: "Created at", key: "created_at" },
-//   { title: "Reason for Voucher", key: "reason", sortable: false },
-//   { title: "VMs", key: "vms" },
-//   { title: "Public IPs", key: "public_ips" },
-//   { title: "Actions", key: "actions", sortable: false },
-// ]);
-
-// const usersHeaders = ref([
-//   { title: "No", key: "id", sortable: false },
-//   { title: "Name", key: "name", sortable: false },
-//   { title: "VMs", key: "vms", sortable: false },
-//   { title: "IPs", key: "public_ips", sortable: false },
-//   { title: "Actions", key: "actions", sortable: false },
-// ]);
-
-// const form = ref(null);
-
+const vouchers = ref([]);
+const pendingVouchers = ref([]);
+const approveAllCount = ref(0);
+const userInfo = ref(null);
 const users = ref([]);
 const toast = ref();
 const tabs = ref([
@@ -524,58 +90,20 @@ const tabs = ref([
   { name: "Users History", route: `/admin/history` },
 ]);
 const activeTab = ref(tabs.value[0]);
-
-// const loading = ref(false);
+const loading = ref(false);
 const usedResources = ref(0);
 const deployedResources = ref(0);
 const balance = ref(0);
 const usedIPs = ref(0);
 const reservedIPs = ref(0);
-// const itemsPerPage = ref(5);
-const dialog = ref(false);
 const announcementDialog = ref(false);
 // const showUserInfo = ref(false);
-const vms = ref(1);
-const ips = ref(0);
-const length = ref(3);
-// const message = ref(null);
-// const voucher = ref(null);
 const subject = ref(null);
 const announcement = ref(null);
 const nextLaunchDialog = ref(false);
 const store = useUserStore();
 const { isNextLaunchEnabled } = storeToRefs(store);
 nextLaunchDialog.value = localStorage.getItem("nextlaunchadmin") == "true";
-
-// const requiredRules = ref([
-//   (value) => {
-//     if (value === "") return "Field is required";
-//     return true;
-//   },
-// ]);
-
-// const setAdmin = (user, admin) => {
-//   userService
-//     .setAdmin(user.email, admin)
-//     .then((response) => {
-//       toast.value.toast(response.data.msg, "#388E3C");
-
-//       userService.getUser().then((response) => {
-//         const current_user = response.data.data.user;
-//         if (!current_user.admin) {
-//           router.push({
-//             name: "Home",
-//           });
-//         }
-//       });
-
-//       getUsers();
-//     })
-//     .catch((response) => {
-//       const { err } = response.response.data;
-//       toast.value.toast(err, "#FF5252");
-//     });
-// };
 
 // const openUserInfo = (user) => {
 //   showUserInfo.value = true;
@@ -627,19 +155,51 @@ const getBalance = () => {
     });
 };
 
-if (localStorage.getItem("token")) {
-  setInterval(() => {
-    getBalance();
-  }, 30 * 1000);
+async function getVouchers() {
+  loading.value = true;
+
+  try {
+    const response = await userService.getVouchers();
+    const { data } = response.data;
+
+    await updateVouchers(data);
+
+    vouchers.value = data.filter(
+      (voucher) => voucher.approved || voucher.rejected
+    );
+    pendingVouchers.value = data.filter(
+      (voucher) => !voucher.approved && !voucher.rejected
+    );
+  } catch (error) {
+    const { err } = error.response.data;
+    toast.value.toast(err, "#FF5252");
+  } finally {
+    loading.value = false;
+  }
 }
 
-watch(dialog, (val) => {
-  if (val) {
-    vms.value = 1;
-    ips.value = 0;
-    length.value = 3;
-  }
-});
+async function updateVouchers(data) {
+  const updatePromises = data.map(async (voucher) => {
+    await new Promise((resolve) => setTimeout(resolve, 10));
+
+    if (voucher.approved && voucher.rejected) {
+      approveAllCount.value++;
+    }
+
+    if (users.value && voucher.user_id) {
+      userInfo.value = users.value.find((user) => user.ID === voucher.user_id);
+
+      if (userInfo.value && voucher.user_id === userInfo.value.ID) {
+        Object.assign(voucher, {
+          email: userInfo.value.email,
+          name: userInfo.value.first_name,
+        });
+      }
+    }
+  });
+
+  await Promise.all(updatePromises);
+}
 
 watch(announcementDialog, (val) => {
   if (val) {
@@ -692,7 +252,8 @@ function setNextLaunch() {
     .then((response) => {
       const { msg } = response.data;
       toast.value.toast(msg, "#388E3C");
-      store.getNextLaunch();
+      store.next_launch_admin = false;
+      router.push("/nextlaunch");
     })
     .catch((response) => {
       const { err } = response.response.data;
@@ -700,62 +261,10 @@ function setNextLaunch() {
     });
 }
 
-// const resetVoucher = () => {
-//   message.value = null;
-//   voucher.value = null;
-// };
-
-// const dateIsNull = (date) => {
-//   return (
-//     new Date(date).toLocaleString() == new Date("0001-01-01").toLocaleString()
-//   );
-// };
-
-// const stringDate = (date) => {
-//   return new Date(date).toLocaleString();
-// };
-
 onMounted(() => {
-  let token = localStorage.getItem("token");
-  if (token) {
-    getUsers();
-    // getVouchers();
-    getBalance();
-    getDeploymentsCount();
-  }
+  getUsers();
+  getVouchers();
+  getBalance();
+  getDeploymentsCount();
 });
-// beforeRouteEnter(to, from, next) {
-//   userService
-//     .getUser()
-//     .then((response) => {
-//       const { user } = response.data.data;
-//       const isAdmin = user.admin;
-//       if (isAdmin) {
-//         next();
-//       } else {
-//         next("/home");
-//       }
-//     })
-//     .catch((err) => {
-//       console.log(err);
-//     });
-// },
 </script>
-
-<!-- <style>
-.resources {
-  margin-top: 0.5rem;
-}
-
-.balance {
-  border: 3px solid #217dbb;
-}
-
-td {
-  line-height: 1em;
-}
-
-.shadow {
-  box-shadow: 0px 2px 4px 2px rgba(0, 0, 0, 0.1);
-}
-</style> -->
